# docs/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AutoCare Hub API
  version: 1.0.0
  description: >
    API para gerenciamento de clientes, veículos, serviços, peças, ordens de
    serviço, orçamentos, aprovações e acompanhamento de status de uma oficina
    mecânica.
servers:
  - url: http://localhost:8080
    description: Ambiente local
security:
  - bearerAuth: [ ]
...
paths:
...
  /api/v1/customers:
    post:
      tags:
        - Customers
      summary: Cria um cliente
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
            examples:
              pessoaFisica:
                summary: Cliente pessoa física com CPF válido
                value:
                  name: Yasmin Barcelos
                  document: '52998224725'
                  phone: '11987654321'
                  email: yasmin.cliente@example.com
                  address:
                    street: Avenida Paulista
                    number: '1000'
                    complement: Sala 10
                    neighborhood: Bela Vista
                    city: São Paulo
                    state: SP
                    zipCode: 01310-100
              pessoaJuridica:
                summary: Cliente empresa com CNPJ válido
                value:
                  name: Oficina Parceira Ltda
                  document: '11222333000181'
                  phone: '1133332222'
                  email: contato@oficinaparceira.com.br
                  address:
                    street: Rua Vergueiro
                    number: '1500'
                    neighborhood: Vila Mariana
                    city: São Paulo
                    state: SP
                    zipCode: 04101-000
      responses:
        '201':
          description: Cliente criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags:
        - Customers
      summary: Lista clientes
      operationId: listCustomers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: active
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Clientes encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/customers/{customerId}:
    get:
      tags:
        - Customers
      summary: Busca um cliente por id
      operationId: getCustomerById
      parameters:
        - $ref: '#/components/parameters/CustomerIdPathParam'
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Customers
      summary: Atualiza um cliente
      operationId: updateCustomer
      parameters:
        - $ref: '#/components/parameters/CustomerIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
            examples:
              atualizarCliente:
                summary: Atualiza cliente mantendo ativo
                value:
                  name: Mariana Costa
                  document: '12345678909'
                  phone: '11988887777'
                  email: mariana.costa@example.com
                  address:
                    street: Rua das Oficinas
                    number: '120'
                    neighborhood: Centro
                    city: São Paulo
                    state: SP
                    zipCode: 01001-000
                  active: true
      responses:
        '200':
          description: Cliente atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Customers
      summary: Remove logicamente um cliente
      operationId: deleteCustomer
      parameters:
        - $ref: '#/components/parameters/CustomerIdPathParam'
      responses:
        '204':
          description: Cliente removido
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/demo-leads:
    post:
      tags:
        - Demo Leads
      summary: Registra interesse em parceria
      description: >
        Endpoint publico usado pela pagina de parceiros para registrar oficinas ou lojas interessadas em demonstração.
      operationId: registerDemoLead
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DemoLeadRequest'
            example:
              contactName: Ana Oficina
              companyName: Oficina Central AutoCare
              demoProfile: workshop
              email: ana@example.com
              phone: '11999999999'
              cnpj: '11.222.333/0001-81'
              city: Sao Paulo
              message: Quero conhecer a plataforma.
      responses:
        '201':
          description: Interesse registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemoLeadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Demo Leads
      summary: Lista interessados em parceria
      description: >
        Endpoint administrativo para Admin Master acompanhar interessados, contatos e parceiros potenciais.
        Requer Bearer Token JWT de administrador.
      operationId: listDemoLeads
      responses:
        '200':
          description: Interessados encontrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DemoLeadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CustomerIdPathParam:
      name: customerId
      in: path
      required: true
      example: 10000000-0000-0000-0000-000000000001
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Requisição invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 400
            error: Bad Request
            message: Invalid request body
            path: /api/v1/service-orders
            details:
              - customerDocument: must not be blank
    Unauthorized:
      description: Credenciais ausentes ou invalidas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 401
            error: Unauthorized
            message: Missing or invalid credentials
            path: /api/v1/customers
            details: [ ]
    Forbidden:
      description: Usuário autenticado sem permissão para acessar o recurso, ou token externo ausente/invalido nos webhooks simulados
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 403
            error: Forbidden
            message: Access denied
            path: /api/v1/customers
            details: [ ]
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 404
            error: Not Found
            message: Resource not found
            path: /api/v1/customers/00000000-0000-0000-0000-000000000000
            details: [ ]
...
    InternalServerError:
      description: Erro interno inesperado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 500
            error: Internal Server Error
            message: Unexpected error
            path: /api/v1/service-orders
            details: [ ]
  schemas:
...
    Address:
      type: object
      required:
        - street
        - number
        - neighborhood
        - city
        - state
        - zipCode
      properties:
        street:
          type: string
          minLength: 2
          maxLength: 120
          example: Avenida Paulista
        number:
          type: string
          minLength: 1
          maxLength: 20
          example: '1000'
        complement:
          type: string
          maxLength: 80
          example: Sala 10
        neighborhood:
          type: string
          minLength: 2
          maxLength: 80
          example: Bela Vista
        city:
          type: string
          minLength: 2
          maxLength: 80
          example: Sao Paulo
        state:
          type: string
          minLength: 2
          maxLength: 2
          example: SP
        zipCode:
          type: string
          minLength: 8
          maxLength: 9
          pattern: "^\\d{5}-?\\d{3}$"
          example: 01310-100
    CreateCustomerRequest:
      type: object
      required:
        - name
        - document
        - phone
        - email
        - address
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        document:
          type: string
          minLength: 11
          maxLength: 18
          pattern: "^\\d{11}$|^\\d{14}$|^\\d{3}\\.\\d{3}\\.\\d{3}-\\d{2}$|^\\d{2}\\.\\d{3}\\.\\d{3}/\\d{4}-\\d{2}$"
          description: CPF ou CNPJ com ou sem formatação.
        phone:
          type: string
          minLength: 10
          maxLength: 20
        email:
          type: string
          format: email
          maxLength: 120
        address:
          $ref: '#/components/schemas/Address'
    UpdateCustomerRequest:
      allOf:
        - $ref: '#/components/schemas/CreateCustomerRequest'
        - type: object
          required:
            - active
          properties:
            active:
              type: boolean
    CustomerResponse:
      type: object
      required:
        - id
        - name
        - document
        - phone
        - email
        - address
        - active
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        document:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        address:
          $ref: '#/components/schemas/Address'
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CustomerListResponse:
      type: object
      required:
        - items
        - page
        - size
        - totalItems
        - totalPages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CustomerResponse'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalItems:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    DemoLeadRequest:
      type: object
      required:
        - contactName
        - companyName
        - demoProfile
        - email
        - phone
        - cnpj
      properties:
        contactName:
          type: string
          minLength: 1
          maxLength: 120
        companyName:
          type: string
          minLength: 1
          maxLength: 120
        demoProfile:
          type: string
          enum:
            - workshop
            - partsStore
        email:
          type: string
          format: email
          maxLength: 160
        phone:
          type: string
          minLength: 8
          maxLength: 30
        cnpj:
          type: string
          minLength: 6
          maxLength: 40
          pattern: '^[A-Za-z0-9./-]+$'
        city:
          type: string
          maxLength: 120
        message:
          type: string
          maxLength: 500
    DemoLeadResponse:
      allOf:
        - $ref: '#/components/schemas/DemoLeadRequest'
        - type: object
          required:
            - id
            - createdAt
          properties:
            id:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          minimum: 100
          maximum: 599
        error:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        path:
          type: string
          minLength: 1
        details:
          type: array
          items:
            type: string